move navigate call from API hook to app buttons
